+++ prompts/loop_breaker.py
@@ -1,0 +1,15 @@
+def create_loop_breaker_prompt(loop_description: str) -> str:
+    return f"""
+[SYSTEM NOTICE: Loop Detected]
+
+The system has detected that you may be stuck in a repetitive pattern:
+{loop_description}
+
+To break out of this loop, please:
+1. Stop and reflect on what you're trying to accomplish
+2. Consider a different approach
+3. If the task seems impossible, explain why and ask for clarification
+4. If you're encountering repeated errors, try a fundamentally different solution
+
+Do not repeat the same action again.
+"""
